Uncompressed ordered map tests for ytypes (#873)
* Add support for ending iteration of subtrees in `ForEachDataField`

* Try adding stringer

* install stringer

* fix comment

* fix enum name

* Add `DiffWithAtomic` to support creating Atomic Notifications.

When comparing two GoStructs where a YANG `ordered-by user` list is
involved, it is not possible using a single Notification to represent
the diff such that the Notification can do the following:
1. update a gNMI cache correctly with the atomic notification.
2. unmarshal the notification to the `orig` to get back the `modified`
   GoStruct.

This function achieves this functionality, sharing the underlying
implementation with Diff, where a disclaimer has been added to the doc
comment.

Other helpers have been created as well to be shared with
`TogNMINotifications`.

Other changes:
- `TogNMINotifications` has been changed to follow the same convention of
  putting the non-atomic Notification as the first message instead of
  the last. The reason is that current code MAY be using `notifs[0]` and
  I don't want to break that, and this also seems slightly better from a
  usability perspective.

* Fix bug when IgnoreAdditions option is turned on

* Add tests for PopulateDefaults for ordered maps

* Move TestTogNMINotifications's ordered map tests to exported tests

* Export ConstructJSON ordered map test

* Export remaining ordered map tests

* lint

* Initial tests for uncompressed structs with ordered maps.

* Remove more dead test code

* Uncompressed ordered map test on PopulateDefaults

* Uncompressed GoStruct tests in mostly `ygot` package.

* Simplified `UnmarshalSetRequest` to always unmarshal w/r to the root
  node. This allows atomic Notifications to be deleted correctly.

* Uncompressed ordered map tests for ytypes

* Move TestTogNMINotifications's ordered map tests to exported tests (#867)

* Move TestTogNMINotifications's ordered map tests to exported tests

* lint

* Add doc comments for test helpers.

* add doc comments for ordered map test helpers

* Preserve some backward-compatibilities with certain render functions. (#875)

* Preserve some backward-compatibilities with certain render functions.

Unfortunately Hyrum's law applies to ygot. I am reverting the
normalizing behaviour for internal JSON generation.

* Fix test imports: we can't have exported test importing symbols from … (#876)

* Fix test imports: we can't have exported test importing symbols from unexported test files

* fix

* Enable ordered map generation for YANG `ordered-by user` lists (#874)

* Enable ordered map generation for YANG `ordered-by user` lists

These satisfy the `ygot.OrderedList` interface.

This is a backwards-incompatible change, but is also a bug fix at the
same time.

* GoOrderedList -> GoOrderedMap

* Fix unkeyed list PopulateDefault generation (#877)

* Fix unkeyed list PopulateDefault generation

* Change interface{} -> any

+++ ygot/render_exported_test.go
@@ -366,10 +366,10 @@ func TestConstructJSONOrderedMap(t *testing.T) {
 				"ordered-multikeyed-list": []any{
 					map[string]any{
 						"key1": "foo",
-						"key2": float64(42),
+						"key2": uint64(42),
 						"config": map[string]any{
 							"key1":  "foo",
-							"key2":  float64(42),
+							"key2":  uint64(42),
 							"value": "foo-val",
 						},
 						"state": map[string]any{
@@ -378,10 +378,10 @@ func TestConstructJSONOrderedMap(t *testing.T) {
 					},
 					map[string]any{
 						"key1": "bar",
-						"key2": float64(42),
+						"key2": uint64(42),
 						"config": map[string]any{
 							"key1":  "bar",
-							"key2":  float64(42),
+							"key2":  uint64(42),
 							"value": "bar-val",
 						},
 						"state": map[string]any{
@@ -390,10 +390,10 @@ func TestConstructJSONOrderedMap(t *testing.T) {
 					},
 					map[string]any{
 						"key1": "baz",
-						"key2": float64(84),
+						"key2": uint64(84),
 						"config": map[string]any{
 							"key1":  "baz",
-							"key2":  float64(84),
+							"key2":  uint64(84),
 							"value": "baz-val",
 						},
 						"state": map[string]any{
@@ -457,7 +457,7 @@ func TestConstructJSONOrderedMap(t *testing.T) {
 				"ordered-multikeyed-list": []any{
 					map[string]any{
 						"key1": "foo",
-						"key2": float64(42),
+						"key2": uint64(42),
 						"config": map[string]any{
 							"value": "foo-val",
 						},
@@ -467,7 +467,7 @@ func TestConstructJSONOrderedMap(t *testing.T) {
 					},
 					map[string]any{
 						"key1": "bar",
-						"key2": float64(42),
+						"key2": uint64(42),
 						"state": map[string]any{
 							"value":    "bar-val",
 							"ro-value": "bar-state-val",
@@ -475,7 +475,7 @@ func TestConstructJSONOrderedMap(t *testing.T) {
 					},
 					map[string]any{
 						"key1": "baz",
-						"key2": float64(84),
+						"key2": uint64(84),
 						"config": map[string]any{
 							"value": "baz-val",
 						},
